move controllers with no api presence to server
