+++ po/POTFILES.in
@@ -27,25 +27,19 @@ subiquity/common/serialize.py
 subiquity/common/tests/test_keyboard.py
 subiquity/common/types.py
 subiquity/controller.py
-subiquity/controllers/cmdlist.py
-subiquity/controllers/debconf.py
-subiquity/controllers/error.py
 subiquity/controllers/filesystem.py
 subiquity/controllers/identity.py
 subiquity/controllers/__init__.py
 subiquity/controllers/keyboard.py
 subiquity/controllers/mirror.py
 subiquity/controllers/network.py
-subiquity/controllers/package.py
 subiquity/controllers/proxy.py
 subiquity/controllers/reboot.py
 subiquity/controllers/refresh.py
-subiquity/controllers/reporting.py
 subiquity/controllers/snaplist.py
 subiquity/controllers/ssh.py
 subiquity/controllers/tests/__init__.py
 subiquity/controllers/tests/test_filesystem.py
-subiquity/controllers/userdata.py
 subiquity/controllers/zdev.py
 subiquitycore/async_helpers.py
 subiquitycore/contextlib38.py
@@ -122,9 +116,14 @@ subiquity/models/tests/test_filesystem.py
 subiquity/models/tests/test_mirror.py
 subiquity/models/tests/test_subiquity.py
 subiquity/server/controller.py
+subiquity/server/controllers/cmdlist.py
+subiquity/server/controllers/debconf.py
 subiquity/server/controllers/__init__.py
 subiquity/server/controllers/install.py
 subiquity/server/controllers/locale.py
+subiquity/server/controllers/package.py
+subiquity/server/controllers/reporting.py
+subiquity/server/controllers/userdata.py
 subiquity/server/dryrun.py
 subiquity/server/errors.py
 subiquity/server/__init__.py
